Refactor - 도메인에 따른 패키지 분리
