+++ src/main/java/lotto/domain/store/LottoNumber.java
@@ -1,4 +1,4 @@
-package lotto.domain;
+package lotto.domain.store;
 
 import java.util.List;
 import java.util.Objects;
