+++ src/main/java/lotto/domain/ticket/CustomLottoTicket.java
@@ -1,8 +1,10 @@
-package lotto.domain;
+package lotto.domain.ticket;
+
+import lotto.domain.store.LottoNumber;
 
 import java.util.List;
 
-public class CustomLottoTicket extends LottoTicket{
+public class CustomLottoTicket extends LottoTicket {
 
     public CustomLottoTicket(List<LottoNumber> lottoNumbers) {
         super(lottoNumbers);
