+++ src/test/java/lotto/LottoMoneyTest.java
@@ -1,6 +1,6 @@
 package lotto;
 
-import lotto.domain.LottoMoney;
+import lotto.domain.store.LottoMoney;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.Arguments;
