+++ src/test/java/lotto/LottoStoreTest.java
@@ -1,7 +1,7 @@
 package lotto;
 
-import lotto.domain.LottoMoney;
-import lotto.domain.LottoStore;
+import lotto.domain.store.LottoMoney;
+import lotto.domain.store.LottoStore;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.Arguments;
